+++ parser/openfoodfacts_taxonomy_parser/parser/taxonomy_parser.py
@@ -7,6 +7,7 @@ from typing import Iterator, TypedDict
 import iso639
 from neo4j import GraphDatabase, Session
 
+from .logger import ParserConsoleLogger
 from .exception import DuplicateIDError
 from ..normalizer import normalizing
 
@@ -23,26 +24,6 @@ class NodeData(TypedDict):
     parent_tag: list[str]
     src_position: int | None
     is_before: str
-
-
-class ParserConsoleLogger:
-    def __init__(self):
-        self.parsing_warnings = []  # Stores all warning logs
-        self.parsing_errors = []  # Stores all error logs
-
-    def info(self, msg, *args, **kwargs):
-        """Stores all parsing info logs"""
-        logging.info(msg, *args, **kwargs)
-
-    def warning(self, msg, *args, **kwargs):
-        """Stores all parsing warning logs"""
-        self.parsing_warnings.append(msg % args)
-        logging.warning(msg, *args, **kwargs)
-
-    def error(self, msg, *args, **kwargs):
-        """Stores all parsing error logs"""
-        self.parsing_errors.append(msg % args)
-        logging.error(msg, *args, **kwargs)
 
 
 class Parser:
